refactor: mongoDB 추가로 디렉토리 구조 개선
